Create Swap odd even rows of matrix
Take m as an integer input.

Then take n as an integer input.

Then make a matrix named mat having m rows and n columns which can store only integer type values.

Then take m*n elements as integer inputs and fill the matrix sequentially in a row wise manner starting from the 0th row.

Now transform the matrix as given under.

Swap the elements in the even row with the corresponding elements present in the odd row just after it.

For eg. Swap the elements of the 0th row with the elements of the 1st row.

Swap the elements of the 2nd row with the elements of the 3rd row.

Note: If there are odd number of rows, then the elements in the last row are not swapped.

Then in the end print the final matrix sequentially in a row wise manner starting from the 0th row.

Input Format

In the first line m will be given

In the second line n will be given

Then the m*n elements of the matrix will be given.

Constraints

1<=m<=100

1<=n<=100

-2^31<=mat[i][j]<=2^31-1

Output Format

Print the output accordingly as shown. Make sure to use tab to separate elements within the same line.